Added JDBC logging
Now there is a possibility to enable jdbc logging for connection. All requests are registered using a special driver (https://github.com/sylvainlaurent/JDBC-Performance-Logger) and displayed in a separate docked panel. Registration of requests by default is on, can turn it off in the settings in the general menu by set item "Enable JDBC logging" to false.